chore: rename as per PEP 8
