+++ pyproject.toml
@@ -165,14 +165,14 @@ test = [
 # which executes the function `main` from this package when invoked.
 [project.scripts]  # Optional
 #TODO: ADD ANY SCRIPTS WE WANT TO HAVE
-load-file = "icesat2_tracks.analysis_db.B01_SL_load_single_file:load_file_app"
-make-spectra = "icesat2_tracks.analysis_db.B02_make_spectra_gFT:make_spectra_app"
-plot-spectra = "icesat2_tracks.analysis_db.B03_plot_spectra_ov:plot_spectra"
-make-iowaga-threads-prior = "icesat2_tracks.analysis_db.A02c_IOWAGA_thredds_prior:make_iowaga_threads_prior_app"
-make-b04-angle = "icesat2_tracks.analysis_db.B04_angle:make_b04_angle_app"
-define-angle = "icesat2_tracks.analysis_db.B05_define_angle:define_angle_app"
-correct-separate = "icesat2_tracks.analysis_db.B06_correct_separate_var:correct_separate_app"
-icesat2waves = "icesat2_tracks.app:app"
+load-file = "icesat2waves.analysis_db.B01_SL_load_single_file:load_file_app"
+make-spectra = "icesat2waves.analysis_db.B02_make_spectra_gFT:make_spectra_app"
+plot-spectra = "icesat2waves.analysis_db.B03_plot_spectra_ov:plot_spectra"
+make-iowaga-threads-prior = "icesat2waves.analysis_db.A02c_IOWAGA_thredds_prior:make_iowaga_threads_prior_app"
+make-b04-angle = "icesat2waves.analysis_db.B04_angle:make_b04_angle_app"
+define-angle = "icesat2waves.analysis_db.B05_define_angle:define_angle_app"
+correct-separate = "icesat2waves.analysis_db.B06_correct_separate_var:correct_separate_app"
+icesat2waves = "icesat2waves.app:app"
 
 
 # This is configuration specific to the `setuptools` build backend.
@@ -181,7 +181,7 @@ icesat2waves = "icesat2_tracks.app:app"
 # If there are data files included in your packages that need to be
 # installed, specify them here.
 # TODO: ADD ANY DATA FILES WE WANT TO HAVE
-package-data = {"icesat2_tracks" = ["config/*.json"]}
+package-data = {"icesat2waves" = ["config/*.json"]}
 
 [tool.setuptools.packages.find]
 where = ["src"]
