+++ src/icesat2waves/analysis_db/B02_make_spectra_gFT.py
@@ -20,15 +20,15 @@ import matplotlib
 from matplotlib import pyplot as plt
 import typer
 
-import icesat2_tracks.ICEsat2_SI_tools.generalized_FT as gFT
-import icesat2_tracks.ICEsat2_SI_tools.iotools as io
-import icesat2_tracks.ICEsat2_SI_tools.spectral_estimates as spec
-import icesat2_tracks.local_modules.m_general_ph3 as M
-import icesat2_tracks.local_modules.m_spectrum_ph3 as spicke_remover
-import icesat2_tracks.local_modules.m_tools_ph3 as MT
-from icesat2_tracks.config.IceSAT2_startup import mconfig
+import icesat2waves.tools.generalized_FT as gFT
+import icesat2waves.tools.iotools as io
+import icesat2waves.tools.spectral_estimates as spec
+import icesat2waves.local_modules.m_general_ph3 as M
+import icesat2waves.local_modules.m_spectrum_ph3 as spicke_remover
+import icesat2waves.local_modules.m_tools_ph3 as MT
+from icesat2waves.config.startup import mconfig
 
-from icesat2_tracks.clitools import (
+from icesat2waves.clitools import (
     echo,
     validate_batch_key,
     validate_output_dir,
