+++ src/icesat2waves/analysis_db/B03_plot_spectra_ov.py
@@ -11,18 +11,18 @@ import xarray as xr
 from matplotlib.gridspec import GridSpec
 import typer
 
-import icesat2_tracks.ICEsat2_SI_tools.iotools as io
-import icesat2_tracks.ICEsat2_SI_tools.generalized_FT as gFT
-import icesat2_tracks.local_modules.m_tools_ph3 as MT
-from icesat2_tracks.local_modules import m_general_ph3 as M
-from icesat2_tracks.config.IceSAT2_startup import (
+import icesat2waves.tools.iotools as io
+import icesat2waves.tools.generalized_FT as gFT
+import icesat2waves.local_modules.m_tools_ph3 as MT
+from icesat2waves.local_modules import m_general_ph3 as M
+from icesat2waves.config.startup import (
     mconfig,
     color_schemes,
     plt,
     font_for_print,
 )
 
-from icesat2_tracks.clitools import (
+from icesat2waves.clitools import (
     echo,
     validate_batch_key,
     validate_output_dir,
