+++ src/icesat2waves/config/color_def.ai
