+++ src/icesat2waves/config/color_def.json
