+++ src/icesat2waves/local_modules/m_spectrum_ph3.py
@@ -4,7 +4,7 @@ from numpy import pi
 from scipy.special import gammainc
 from scipy import signal
 import matplotlib.pyplot as plt
-from icesat2_tracks.local_modules import m_general_ph3 as M
+from icesat2waves.local_modules import m_general_ph3 as M
 
 try:
     import mkl
