+++ src/icesat2waves/tools/generalized_FT.py
@@ -10,8 +10,8 @@ from scipy.constants import g
 from scipy.signal import detrend
 import lmfit as LM
 
-from icesat2_tracks.ICEsat2_SI_tools import lanczos, spectral_estimates as spec
-import icesat2_tracks.local_modules.JONSWAP_gamma as spectal_models
+from icesat2waves.tools import lanczos, spectral_estimates as spec
+import icesat2waves.local_modules.jonswap_gamma as spectal_models
 
 
 def rebin(data, dk):
